Updated version and comments.
I changed a lot of stuff in the project. I created a for-loop and insert
stuff into the collection. So every time the server start there are data
to be tested. I also added remove() and add() functions for event feed.
Had to do them to test and debug some of my stuff. Lastly, I change all
the $scope into 'this'. I think this is a better style so we know what
controller is being use in what file.